requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frexco-pylogger==0.2.1
hvac==1.0.2
idna==3.4
pika==1.2.0
psycopg2-binary==2.9.5
pyhcl==0.4.4
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
